+++ src/Microsoft.DotNet.Interactive.Jupyter/Connection/IJupyterKernelConnectionOptions.cs
@@ -1,4 +1,7 @@
-﻿using System.Collections.Generic;
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.Parsing;
